requirements.txt
Flask==2.2.5
Werkzeug==2.2.3
opencv-python==4.7.0.72
requests==2.28.2
numpy==1.24.2
Pillow==9.4.0
scikit-image==0.22.0
imutils==0.5.4
gunicorn==20.1.0